requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
albumentations @ file:///home/conda/feedstock_root/build_artifacts/albumentations_1606664784723/work
altair==4.1.0
argon2-cffi @ file:///D:/bld/argon2-cffi_1605217273648/work
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
backcall==0.2.0
base58==2.1.0
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1600454382015/work
blinker==1.4
blis @ file:///D:/bld/cython-blis_1607338345852/work
brotlipy==0.7.0
cachetools==4.2.1
catalogue @ file:///D:/bld/catalogue_1605613788233/work
certifi==2020.12.5
cffi @ file:///D:/bld/cffi_1602537433319/work
chardet @ file:///C:/ci/chardet_1607690654534/work
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cmake==3.18.4.post1
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
cryptography @ file:///C:/ci/cryptography_1607637849569/work
cycler==0.10.0
cymem @ file:///D:/bld/cymem_1607336390769/work
Cython==0.29.21
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1602083700509/work
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.3.0/de_core_news_sm-2.3.0.tar.gz
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.7.1
dlib @ file:///D:/bld/dlib_1617441823598/work
emoji==1.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
fastapi==0.63.0
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imgaug==0.4.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1606126889468/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598865816163/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1599554010055/work
jedi @ file:///C:/ci/jedi_1598353661586/work
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1602551949684/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe @ file:///D:/bld/markupsafe_1602267501600/work
matplotlib==3.3.3
mistune @ file:///D:/bld/mistune_1605115857778/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
murmurhash @ file:///D:/bld/murmurhash_1607334400096/work
mysql-connector==2.2.9
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1602859080374/work
nbconvert @ file:///D:/bld/nbconvert_1605402041690/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1602732862338/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1605195931949/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
notebook @ file:///D:/bld/notebook_1605103957289/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandas==1.1.4
pandas-read-xml==0.0.8
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1603823068645/work
plac==1.1.3
preshed @ file:///D:/bld/preshed_1607337932670/work
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1605543085815/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.14.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydantic==1.8.1
pydeck==0.6.1
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent @ file:///D:/bld/pyrsistent_1605115827510/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2020.4
PyWavelets @ file:///C:/ci/pywavelets_1601658407916/work
pywin32==227
pywinpty @ file:///D:/bld/pywinpty_1602377663724/work
PyYAML==5.3.1
pyzmq==19.0.2
regex @ file:///C:/ci/regex_1606691183008/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn @ file:///C:/ci/scikit-learn_1598377018496/work
scipy @ file:///C:/ci/scipy_1604596260408/work
Send2Trash==1.5.0
Shapely==1.7.1
six @ file:///C:/ci/six_1605187374963/work
smmap==3.0.5
spacy @ file:///D:/bld/spacy_1608283206934/work
srsly @ file:///D:/bld/srsly_1607347939729/work
starlette==0.13.6
streamlit==0.80.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboard-pytorch==0.7.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado @ file:///D:/bld/terminado_1605117138129/work
testpath==0.4.4
thinc @ file:///D:/bld/thinc_1608283219407/work
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2020.10.1
toml==0.10.2
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
torch==1.7.0
torchaudio==0.7.0
torchtext==0.8.0
torchvision==0.8.1
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
tzlocal==2.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
uvicorn==0.13.4
validators==0.18.2
wasabi @ file:///home/conda/feedstock_root/build_artifacts/wasabi_1600272362626/work
watchdog==2.0.2
watchgod==0.7
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension @ file:///D:/bld/widgetsnbextension_1605475695976/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1603668650351/work
